requirements.txt
QtPy>=2.4.0
PySide6>=6.6.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0